refactor: 抽 vision.rs 模块，VL 逻辑独立（行为不变）(#1468 Bug 4.1)
- VL 预处理代码（VL_IMAGE_LIMIT/VL_SINGLE_TIMEOUT 常量 + estimate_item_tokens/
  items_within_vl_window/collect_input_text/extract_image_url/describe_image_with_vl/
  analyze_images_with_vl/apply_vl_with_fallback）从 protocol_proxy.rs 移到 vision.rs
- 公开 API 不变：analyze_images_with_vl / apply_vl_with_fallback 仍 pub，签名不变
- protocol_proxy.rs:531 调用改 crate::vision::apply_vl_with_fallback
- 测试 analyze_images_with_vl/apply_vl_with_fallback 改从 codex_plus_core::vision 导入
- 纯移动，84 测试全绿，行为不变（为后续两层缓存/并发/重试/超时迭代做铺垫）
