+++ crates/codex-plus-core/src/lib.rs
@@ -37,6 +37,7 @@ pub mod update;
 pub mod upstream_worktree;
 pub mod user_scripts;
 pub mod version;
+pub mod vision;
 pub mod watcher;
 #[cfg(windows)]
 mod windows_integration;
